% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGWAS.R
\name{EGWAS}
\alias{EGWAS}
\title{Ensemble-like Genome Wide Associated Analysis Strategy}
\usage{
EGWAS(
  y = NULL,
  geno = NULL,
  map = NULL,
  CV = NULL,
  pvalue = NULL,
  pip = NULL,
  bin.size = 0,
  permu.num = 1000,
  pv.threshold=0.05,
  pip.threshold=0.05,
	permu.threshold=0.05,
  Kinship = NULL,
  LD.threshold = 0.7,
  ncpus = 1,
  verbose = TRUE
)
}
\arguments{
\item{y}{vector of phenotype, NA is NOT allowed}

\item{geno}{numeric genotype, row for individual and column for SNP}

\item{map}{map data, 3 columns for SNPs name, Chromosome, Position}

\item{CV}{additional fixed effects or covariates}

\item{pvalue}{p values of GWAS methods, ncol equal to nmethod}

\item{pip}{posterior inclusive probability (PIP) derived from Bayesian method}

\item{bin.size}{size of bin in union-intersection step}

\item{permu.num}{number of permutation test}

\item{pv.threshold}{significant level, Bonferroni}

\item{pip.threshold}{significant level, PIP}

\item{permu.threshold}{significant level, Permutation}

\item{Kinship}{genetic relationship matrix}

\item{LD.threshold}{correlation threshold for cluster}

\item{ncpus}{number of cpus used for parallel}

\item{verbose}{whether to print detail.}
}
\value{
a list of pseudo QTNs detected by base methods and EGWAS
}
\description{
a post GWAS strategy aims to control false positive rate inspired by ensemble learning method.
}
\examples{
\donttest{
eg <- EGWAS(y=phe, geno=geno, map=map,CV=NULL, pvalue=GWAS, pip=NULL,
				  bin.size=0, permu.num=1000,pv.threshold=0.05,pip.threshold=0.05,
				  permu.threshold=0.01,Kinship=NULL,LD.threshold=0.7,
				  ncpus=10,verbose=TRUE)
}
}
