% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGWAS.R
\name{EGWAS.SCAN}
\alias{EGWAS.SCAN}
\title{Title}
\usage{
EGWAS.SCAN(
  y = NULL,
  geno = NULL,
  map = NULL,
  CV = NULL,
  EGWAS.QTN = NULL,
  Kinship = NULL,
  ncpus = 1
)
}
\arguments{
\item{y}{vector of phenotype, NA is NOT allowed}

\item{geno}{numeric genotype, row for individual and column for SNP}

\item{map}{map data, 3 columns for SNPs name, Chromosome, Position}

\item{CV}{additional fixed effects or covariates}

\item{EGWAS.QTN}{index of pseudo QTNs detected by E-GWAS}

\item{Kinship}{genetic relationship matrix}

\item{ncpus}{number of cpus used for parallel}
}
\value{
a data.frame contain map, beta, se, pvalues
}
\description{
MLM scan based on pseudo QTNs (as fix effects) from EGWAS
}
\examples{
\donttest{
myscan <- EGWAS.SCAN(y=phe, geno=geno, map=map,CV=NULL,
  EGWAS.QTN=myEGWAS$EGWAS,Kinship=G,ncpus=10)
}
}
