% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGWAS.R
\name{EGWAS}
\alias{EGWAS}
\title{Ensemble-like Genome Wide Associated Analysis Strategy}
\usage{
EGWAS(
  y = NULL,
  geno = NULL,
  map = NULL,
  CV = NULL,
  GWAS.results = NULL,
  Bayesian.method = c(),
  Bin.size = 0,
  permu.num = 1000,
  p.threshold = c(0.05, 0.05, 0.05),
  Kinship = NULL,
  LD.threshold = 0.7,
  ncpus = 1,
  verbose = TRUE
)
}
\arguments{
\item{y}{vector of phenotype, NA is NOT allowed}

\item{geno}{numeric genotype, row for individual and column for SNP}

\item{map}{map data, 3 columns for SNPs name, Chromosome, Position}

\item{CV}{additional fixed effects or covariates}

\item{GWAS.results}{p values of GWAS methods, ncol equal to nmethod}

\item{Bayesian.method}{column for Bayesian method in GWAS.results}

\item{Bin.size}{size of bin in union-intersection step}

\item{permu.num}{number of permutation test}

\item{p.threshold}{significant level, Bonferroni/PPI/Permutation}

\item{Kinship}{genetic relationship matrix}

\item{LD.threshold}{correlation threshold for cluster}

\item{ncpus}{number of cpus used for parallel}

\item{verbose}{whether to print detail.}
}
\value{
a list of pseudo QTNs detected by base methods and EGWAS
}
\description{
a post GWAS strategy aims to control false positive rate inspired by ensemble learning method.
}
\examples{
\donttest{
myEGWAS <- EGWAS(y=phe, geno=geno, map=map,CV=NULL,
  GWAS.results=GWAS,Bayesian.method=c(1),
  Bin.size=0, permu.num=1000,p.threshold=C(.05,.05,.05),
  Kinship=G,LD.threshold=0.7,ncpus=10,verbose=TRUE)
}
}
