% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGWAS.R
\name{EGWAS}
\alias{EGWAS}
\title{Ensemble-like Genome Wide Associated Analysis Strategy}
\usage{
EGWAS(
  y = NULL,
  geno = NULL,
  map = NULL,
  CV = NULL,
  pvalue = NULL,
  pip = NULL,
  bin.size = 0,
  permu.num = 1000,
  pv.threshold=0.05,
  pip.threshold=0.05,
	permu.threshold=0.05,
  Kinship = NULL,
  LD.threshold = 0.7,
  ncpus = 1,
  verbose = TRUE
)
}
\arguments{
\item{y}{n * 1 vector, phenotypic value, NA is NOT allowed, n is the number of sample individuals.}

\item{geno}{m * n numeric matrix, genotype data, m is the number of markers, n is the number of sample individuals.}

\item{map}{m * 3 data.frame, m is the number of markers, 3 columns are SNP name, Chromosome, Position.}

\item{CV}{n * q co-variate matrix, n is the number of sample individuals, and q is the number of additional co-variate (except the group mean). Quantitative variables are entered in numeric format and discrete / categorical variables are entered in factor or character format. The default value is null, that is, only the mean as co-variate.}

\item{pvalue}{n * i matrix, the p-value generated by the GWAS method based on hypothesis testing, n is the number of sample individuals, and i is the number of methods.}

\item{pip}{n * j matrix, the posterior inclusion probability (pip) generated based on the Bayesian methods, n is the number of sample individuals and j is the number of methods.}

\item{bin.size}{Numeric, which determine that the number of bp between 2 significant SNPs from different method regarded as the same in the pairwise intersection step. The default value is 0.}

\item{permu.num}{Numeric, the number of replicates during the permutation test, the default value is 1000.}

\item{pv.threshold}{Numeric, Bonferroni significance threshold for "pvalue", with a default value of 0.05, where SNPs with " pvalue' < 0.05 / m were detected as significant loci.}

\item{pip.threshold}{Numeric, significance threshold for "pip", with a default value of 0.05, where SNPs with (1- "pip") < 0.05 were detected as significant loci.}

\item{permu.threshold}{Numeric, significance threshold for permutation test, with a default value of 0.05}

\item{Kinship}{n * n numeric matrix, the individual genomic relationship matrix, n is the number of sample individuals, the default value is NULL, which the "Kinship" is calculated based on "genotype" by the VanRaden method.}

\item{LD.threshold}{Numeric, correlation threshold, with a default value of 0.7. The candidate SNPs were clustered based on the threshold, leaving only the most significant candidate SNP in each cluster.}

\item{ncpus}{Numeric, the number of threads used, and the default is 1. Make sure that the computer has enough memory to process the input data when used parallel computing, especially in windows systems.}

\item{verbose}{Logical, whether to print the running procedure, which is TRUE by default.}
}
\details{
Genome-wide association studies (GWAS) are an effective method to explore genotype-phenotype associations in humans and animals and plants, making a significant contribution to exploring the genetic mechanisms of complex traits. The researchers have proposed multiple GWAS methods based on different statistical or genetic hypotheses. However, due to the diversity and complexity of trait genetic structure, no single method can be applied to all traits. And for many traits, the significant loci detected using different GWAS methods are not entirely consistent. Therefore, GWAS strategies combining different methods are increasingly applied, but this combined use is only a simple merger of the results, and the lack of statistical analysis of the combined results, does not a real integrative analysis. Inspired by the integrated learning method in the field of machine learning, we put forward a kind of integrated GWAS strategy (EGWAS), the strategy from different single GWAS method of significant site results pair intersection again, combined with linkage disequilibrium screening and replacement test, which can not reduce the basis of strictly control the false positive.
}
\value{
A list of length 3. $Base.signal is the significant SNPs detected by each GWAS method; $EGWAS is the candidate detected by EGWAS; $Specific is the specific significant SNPs in each method.
}
\description{
A post GWAS strategy aims to control false positive rate inspired by ensemble learning method.
}
\examples{
\donttest{
eg <- EGWAS(y=phe, geno=geno, map=map,CV=NULL, pvalue=GWAS, pip=NULL,
				  bin.size=0, permu.num=1000,pv.threshold=0.05,pip.threshold=0.05,
				  permu.threshold=0.01,Kinship=NULL,LD.threshold=0.7,
				  ncpus=10,verbose=TRUE)
}
}
